Brew

+++ init.sh
@@ -1,6 +1,43 @@
 #! /usr/bin/zsh
 
+# Variables
+OK="link updated"
+KO="link skipped"
+
 # setup links for local dotfiles
-ln -s ~/.dotfiles-mac/.gitconfig ~/.gitconfig
-ln -s ~/.dotfiles-mac/.zshrc ~/.zshrc
-ln -s ~/.dotfiles-mac/.aws ~/.aws
+
+# Brewfile
+if [ -e ~/.Brewfile ]
+then
+	echo $KO
+else
+	ln -s ~/.dotfiles-mac/Brewfile ~/.Brewfile
+	echo $OK
+fi
+
+# AWS
+if [ -e ~/.aws ]
+then
+	echo $KO
+else
+	ln -s ~/.dotfiles-mac/.aws ~/.aws
+	echo $OK
+fi
+
+# Git
+if [ -e ~/.gitconfig ]
+then
+	echo $KO
+else
+	ln -s ~/.dotfiles-mac/.gitconfig ~/.gitconfig
+	echo $OK
+fi
+
+# zsh
+if [ -e ~/.zshrc ]
+then
+	echo $KO
+else
+	ln -s ~/.dotfiles-mac/.zshrc ~/.zshrc
+	echo $OK
+fi
